Include pcr7 data from updating shim to include GUID
Bootkit now includes the uki guids when computing the vendor_db.esl
for shim. Thus we need to include the uki guids when computing the
pcr7 values for a keyset resulting in new pcr7 data for snakeoil.
This commit adds new pcr7 values and policy into pcr7data.
The original ones can possibly be deprecated.

Signed-off-by: Joy Latten <jlatten@cisco.com>